# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestão de ordens de serviço para oficina mecânica",
        "title": "AutoShop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte AutoShop",
            "email": "suporte@autoshop.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "",
    "basePath": "/",
    "paths": {
...
        "/parts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todas as peças e insumos cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Listar peças",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Part"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra uma nova peça ou insumo com estoque inicial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Criar peça",
                "parameters": [
                    {
                        "description": "Dados da peça",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePartRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Part"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/parts/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna peças com estoque abaixo ou igual ao mínimo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Peças com estoque baixo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Part"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/parts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma peça pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Buscar peça",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Part"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de uma peça existente (sem alterar estoque)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Atualizar peça",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Part"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove uma peça do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Deletar peça",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/parts/{id}/stock": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adiciona ou remove quantidade do estoque (use negativo para saída)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parts"
                ],
                "summary": "Ajustar estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ajuste de estoque",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdjustStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Part"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Part": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "description": "oleo, filtro oleo, pastiha freio, parafuso, correia dentada",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                },
                "unit": {
                    "description": "unidade, litro, metro, par",
                    "type": "string"
                }
            }
        },
...
        "dto.AdjustStockRequest": {
            "type": "object",
            "required": [
                "quantity",
                "reason"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "description": "entrada, saída, ajuste",
                    "type": "string"
                }
            }
        },
...
        "dto.CreatePartRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "unit"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePartRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "unit"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite: Bearer {seu_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}